some fixes on the contour plotting to avoid interminable times on weird contours

+++ _AlgorithmsPlugins/Projections/qcontour.h
@@ -7,10 +7,10 @@
 // implementation of the CRaster class for contour lines creation
 class ValueMap:public CRaster
 {
-    float *values;
+    double *values;
 public:
     int w, h;
-    ValueMap(float *values=0, int w=0, int h=0):values(values), w(w), h(h){}
+    ValueMap(double *values=0, int w=0, int h=0):values(values), w(w), h(h){}
     double value(double x,double y){return values && w ? values[int(y)*w + int(x)] : 0;}
     SPoint upper_bound(){return SPoint(w-1,h-1);}
     SPoint lower_bound(){return SPoint(0,0);}
@@ -23,11 +23,11 @@ public:
 class QContour
 {
     ValueMap valueMap;
-    float vmin, vmax;
-    float meanValue(int xStart, int xEnd, int yStart, int yEnd);
-    float meanValue(QRect rect);
+    double vmin, vmax;
+    double meanValue(int xStart, int xEnd, int yStart, int yEnd);
+    double meanValue(QRect rect);
 public:
-    QContour(float *values, int w, int h);
+    QContour(double *values, int w, int h);
     void Paint(QPainter &painter, int levels);
 };
